FIX Periodic dtypes in tests

+++ skmine/periodic/tests/test_cycles.py
@@ -175,9 +175,9 @@ def test_evaluate():
     cycles, residuals = evaluate(minutes, cands)
     assert residuals.size == 0
     assert cycles.dtypes.to_dict() == {
-        "start": np.dtype("int64"),
-        "length": np.dtype("int64"),
-        "period": np.dtype("int64"),
+        "start": np.dtype("int"),
+        "length": np.dtype("int"),
+        "period": np.dtype("int"),
         "dE": np.dtype("object"),
     }
     assert cycles.length.tolist() == [6, 3, 3]
